services/core-service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
pydantic-settings
loguru
greenlet
aiofiles
pyyaml
python-multipart
motor
pydantic[email]
httpx
passlib
argon2_cffi
